cards v0.1.0 Cards

Provides methods for creating and handling a deck of cards

Example usage

iex> deck = Cards.create_deck
iex> {hand, _deck} = Cards.deal(deck, 1)
iex> hand
["Ace of Spades"]

Functions

contains?(deck, card)

Checks if a deck contains a card

## Examples

iex> deck = Cards.create_deck
  iex> Cards.contains?(deck, "Ace of Spades")
  true

create_deck()

Creates an array that represents a deck of cards

## Examples

iex> Cards.create_deck
["Ace of Spades", "Two of Spades", "Three of Spades", "Four of Spades",
"Five of Spades", "Ace of Clubs", "Two of Clubs", "Three of Clubs",
"Four of Clubs", "Five of Clubs", "Ace of Hearts", "Two of Hearts",
"Three of Hearts", "Four of Hearts", "Five of Hearts", "Ace of Diamonds",
"Two of Diamonds", "Three of Diamonds", "Four of Diamonds", "Five of Diamonds"]

create_hand(hand_size)

Creates a deck, shuffles it, and deals one hand of hand_size

deal(deck, hand_size)

Takes hand_size cards out of the deck and returns two arrays:

one with hand_size cards

another one with the remaining deck

Examples

iex> deck = Cards.create_deck
iex> {hand, _deck} = Cards.deal(deck, 1)
iex> hand
["Ace of Spades"]

load(filename)

Loads a binary saved deck from disk.

save(deck, filename)

Saves a deck to disk in binary format

shuffle(deck)

Shuffles a deck randomly